04 auth

+++ service/contactsService.js
@@ -1,27 +1,50 @@
 const { Contact } = require('../models/contactsModel');
 
-const getCnt = async () => {
-    return Contact.find();
+const getCnt = async (userId, query) => {
+    const { page = 1, limit = 20, favorite, name, email } = query;
+    const skip = (page - 1) * limit;
+
+    const filter = {};
+    if (userId) {
+        filter.owner = userId;
+    }
+    if (name) {
+        filter.name = name;
+    }
+    if (email) {
+        filter.email = email;
+    }
+    if (favorite) {
+        filter.favorite = favorite;
+    }
+
+    return Contact.find(
+        filter, "", { skip, limit: +limit }
+    ).populate('owner', '_id username email subscription');
 }
 
-const getCntById = async id => {
-    return Contact.findById({ _id: id });
+const getCntById = async (id, userId) => {
+    return Contact.findById({ _id: id, owner: userId });
 }
 
 const addCnt = async body => {
     return Contact.create(body);
 }
 
-const updateCnt = async (id, body) => {
-    return Contact.findByIdAndUpdate({ _id: id }, body, { new: true });
+const updateCnt = async ({ _id, contactId, ...body }) => {
+    return Contact.findByIdAndUpdate({ _id: contactId, owner: _id, body, new: true });
 }
 
-const updateStatusCnt = async (id, body) => {
-    return Contact.findByIdAndUpdate({ _id: id }, { $set: { favorite: body } }, { new: true });
+const updateStatusCnt = async (id, userId, body) => {
+    return Contact.findByIdAndUpdate(
+        { _id: id, owner: userId },
+        { $set: { favorite: body } },
+        { new: true }
+    );
 }
 
-const deleteCnt = async id => {
-    return Contact.findByIdAndRemove({ _id: id });
+const deleteCnt = async (id, userId) => {
+    return Contact.findByIdAndRemove({ _id: id, owner: userId });
 }
 
 module.exports = {
